Re-structure project
