+++ src/components/LabelBook.js
@@ -3,7 +3,7 @@ import {formatCurrency} from '../util/formatNumbers';
 
 
 const MenuBooks = ({ bookSelected, message  }) => {
-	console.log("Render.MenuBooks", this);
+	//console.log("Render.MenuBooks", this);
 	const units = bookSelected.toUpperCase().split('_')
 	return (
 		<span href="#about">1 {units[0]} = {formatCurrency(message)} {units[1]}</span>
