--- src/components/graph/GraphContainer.js
@@ -1,6 +1,0 @@
-import ParentSize from '../util/ParentSize'
-import LineChart from './LineChart'
-
-let GraphContainer = withParentSize(MySuperCoolVxChart);
-
-export default GraphContainer;
