+++ src/containers/FilterBooks.js
@@ -5,7 +5,7 @@ const CHANGE_BOOK = 'CHANGE_BOOK';
 
 
 const mapDispatchToProps = dispatch => {
-	console.log("connect.method.ping FilterBooks", dispatch);
+	//console.log("connect.method.ping FilterBooks", dispatch);
 	return {
 		selectBook : (book, lastBook) => dispatch({
 			type : CHANGE_BOOK,
@@ -18,7 +18,7 @@ const mapDispatchToProps = dispatch => {
 const mapStateToProps = state => {
 	const {bookSelected, books} = state.booksReducer;
 	const {ticker} = state.tradesReducer;
-	console.log("state.MenuBooks", state, bookSelected, books, ticker);
+	//console.log("state.MenuBooks", state, bookSelected, books, ticker);
 	return {bookSelected, books, ticker};
 }
