+++ src/containers/FilterLabelBook.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux'
 import LabelBook from '../components/LabelBook'
 
 const mapStateToProps = state => {
-	console.log("state.MenuBooks", state);
+	//console.log("state.MenuBooks", state);
 	const {bookSelected} = state.booksReducer;
 	const {message} = state.websocketReducer;
 	return {bookSelected, message};
