+++ src/containers/FilterMarketTickers.js
@@ -4,12 +4,12 @@ import MarketTable from '../components/MarketTable'
 const LOAD_ALL_TICKERS = 'LOAD_ALL_TICKERS';
 
 const loadTickers = () => {
-	console.log("connect.method.ping FilterTrades");
+	//console.log("connect.method.ping FilterTrades");
 	return ({ type: LOAD_ALL_TICKERS })
 };
 
 const mapStateToProps = state => {
-	console.log("state.tickersReducer", state);
+	//console.log("state.tickersReducer", state);
 	return state.tickersReducer;
 }
